Changed fundamental sample time from Ts to fpga_Ts
Previously I fixed the spectrum analyzer to display correctly by hard coding the correct sampling frequency but the simulation was running at the DAC sampling rate of 2Gsps rather than the FPGA clock rate. Fixed this and went back to default sampling time on the spectrum analyzer. We really want the sim running at the FPGA clock rate not Ts.